#769 add nested testsuite rule to wahlvorbereitung service
