+++ wls-wahlvorbereitung-service/pom.xml
@@ -37,7 +37,7 @@
         <org.projectlombok.mapstructbinding.version>0.2.0</org.projectlombok.mapstructbinding.version>
         <org.springdoc.version>2.6.0</org.springdoc.version>
         <org.mapstruct.version>1.6.3</org.mapstruct.version>
-        <wls.common.version>1.4.0</wls.common.version>
+        <wls.common.version>1.5.0-SNAPSHOT</wls.common.version>
         <com.tngtech.archunit.version>1.4.0</com.tngtech.archunit.version>
     </properties>
