+++ wls-wahlvorbereitung-service/src/test/java/de/muenchen/oss/wahllokalsystem/wahlvorbereitungservice/rest/urnenwahlvorbereitung/UrnenwahlvorbereitungDTOTest.java
@@ -2,20 +2,14 @@ package de.muenchen.oss.wahllokalsystem.wahlvorbereitungservice.rest.urnenwahlvo
 
 import lombok.val;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 class UrnenwahlvorbereitungDTOTest {
 
-    @Nested
-    class ConstructionTests {
+    @Test
+    void listOfWahlurneIsNotNullOnNullParameter() {
+        val result = new UrnenwahlvorbereitungDTO(null, 0, 0, 0, null);
 
-        @Test
-        void listOfWahlurneIsNotNullOnNullParameter() {
-            val result = new UrnenwahlvorbereitungDTO(null, 0, 0, 0, null);
-
-            Assertions.assertThat(result.urnenAnzahl()).isNotNull();
-        }
+        Assertions.assertThat(result.urnenAnzahl()).isNotNull();
     }
-
 }
