+++ wls-wahlvorbereitung-service/src/test/java/de/muenchen/oss/wahllokalsystem/wahlvorbereitungservice/service/urnenwahlvorbereitung/UrnenwahlvorbereitungModelTest.java
@@ -2,20 +2,14 @@ package de.muenchen.oss.wahllokalsystem.wahlvorbereitungservice.service.urnenwah
 
 import lombok.val;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 class UrnenwahlvorbereitungModelTest {
 
-    @Nested
-    class ConstructionTests {
+    @Test
+    void listOfWahlurneIsNotNullOnNullParameter() {
+        val result = new UrnenwahlvorbereitungModel(null, 0, 0, 0, null);
 
-        @Test
-        void listOfWahlurneIsNotNullOnNullParameter() {
-            val result = new UrnenwahlvorbereitungModel(null, 0, 0, 0, null);
-
-            Assertions.assertThat(result.urnenAnzahl()).isNotNull();
-        }
+        Assertions.assertThat(result.urnenAnzahl()).isNotNull();
     }
-
 }
